// src/app/interfaces/cliente.ts
export interface Cliente {
    idEmpresa: number;
    idCliente: number;
    pessoa: string;
    cpfCnpj: string;
    rgIe: string;
    razaoSocial: string;
    nomeFantasia: string;
    apelido: string;
    cep: string;
    endereco: string;
    numero: string;
    complemento: string;
    bairro: string;
    cidade: string;
    uf: string;
    ibge: string;
    contato: string;
    telefone: string;
    celular: string;
    email: string;
    dadosAdicionais: string;
}